refactor(tui): reuse canonical sleep helper
